requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Werkzeug==2.3.7
SQLAlchemy==2.0.20
python-dotenv==1.0.0
gunicorn==21.2.0
pandas==2.1.0
scikit-learn==1.3.0
matplotlib==3.7.2
python-dateutil==2.8.2
